+++ mobile/android/android-components/buildSrc/src/main/java/Dependencies.kt
@@ -32,7 +32,7 @@ object Versions {
     const val room = "1.1.1"
     const val paging = "1.0.1"
 
-    const val mozilla_appservices = "0.18.0"
+    const val mozilla_appservices = "0.19.0"
     const val servo = "0.0.1.20181017.aa95911"
 }