+++ mobile/android/android-components/docs/changelog.md
@@ -49,6 +49,9 @@ permalink: /changelog/
 * **feature-session**
   * Adds support for the picture-in-picture mode in `PictureInPictureFeature`.
 
+* **browser-storage-sync**
+  * Changed how Rust Places database connections are maintained, based on [new reader/writer APIs](https://github.com/mozilla/application-services/pull/718).
+
 # 0.46.0
 
 * [Commits](https://github.com/mozilla-mobile/android-components/compare/v0.45.0...v0.46.0)